added music (test)

+++ game_sources/include/Entity.h
@@ -7,5 +7,5 @@ class Entity : public sf::Transformable, public sf::Drawable
 {
 public:
 	virtual void ProcessInput() {}
-	virtual void Update(float dt) {}	
+	virtual void Update(size_t dt) {}	
 };
