+++ game_sources/include/Snowflake.h
@@ -11,7 +11,7 @@ public:
 
 	void Initialize(shared_ptr<PlayField> playfield);
 
-	void Update(float dt) override;
+	void Update(size_t dt) override;
 	
 	sf::FloatRect GetBBox() const;
 
@@ -47,7 +47,7 @@ public:
 	void Initialize(shared_ptr<PlayField> playfield);
 
 	void ProcessInput() override;
-	void Update(float dt) override;
+	void Update(size_t dt) override;
 
 	void SetMaximumSnoflakes(size_t number) { m_maxSnowflakeNumber = number; }
 	void TryToCreateSnowflake();
